# server/openapi.yaml
openapi: 3.0.3
info:
  title: GhostTerm Progress Module API
  version: 1.0.0
  description: |
    GhostTerm 进度模块 API 单一来源（OpenAPI 3.0）。
    后端用 ogen 生成 server skeleton，前端用 openapi-typescript / openapi-zod-client 生成类型与运行时 schema。
    每次 schema 变更：改本文件 → make gen → 双方编译错误自动暴露不一致。
servers:
  - url: http://localhost:8080
    description: 本地开发
security:
  - bearerAuth: []

# ======================================================================
# 通用响应包装（v2 强制）：成功 200/201 必须 DataEnvelope，错误必须 ErrorEnvelope
# ======================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - unauthorized
                - permission_denied
                - not_found
                - validation_failed
                - state_machine_invalid_transition
                - mime_not_allowed
                - file_too_large
                - ticket_invalid
                - not_implemented
                - internal
                - super_admin_immutable
                - request_body_too_large
                - service_unavailable
                - invalid_effect
                - permission_not_found
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              nullable: true
...
    FileMetadata:
      type: object
      required: [id, uuid, filename, sizeBytes, mimeType, uploadedBy, uploadedAt]
      properties:
        id: { type: integer, format: int64 }
        uuid: { type: string, format: uuid }
        filename: { type: string }
        sizeBytes: { type: integer, format: int64 }
        mimeType: { type: string }
        uploadedBy: { type: integer, format: int64 }
        uploadedAt: { type: string, format: date-time }
    ProjectFile:
      type: object
      required: [id, projectId, fileId, category, addedAt, file]
      properties:
        id: { type: integer, format: int64 }
        projectId: { type: integer, format: int64 }
        fileId: { type: integer, format: int64 }
        category:
          type: string
          enum: [sample_doc, source_code, wechat_chat]
        addedAt: { type: string, format: date-time }
        remark:
          type: string
          nullable: true
          description: "附加备注：源码上传说明 / 版本号 / 参考样稿描述等（migration 0013）"
        file: { $ref: '#/components/schemas/FileMetadata' }
    ThesisVersion:
      type: object
      required: [id, projectId, fileId, versionNo, uploadedBy, uploadedAt, file]
      properties:
        id: { type: integer, format: int64 }
        projectId: { type: integer, format: int64 }
        fileId: { type: integer, format: int64 }
        versionNo: { type: integer }
        remark: { type: string, nullable: true }
        uploadedBy: { type: integer, format: int64 }
        uploadedAt: { type: string, format: date-time }
        file: { $ref: '#/components/schemas/FileMetadata' }
    ThesisVersionCreateRequest:
      type: object
      required: [fileId]
      properties:
        fileId: { type: integer, format: int64 }
        remark: { type: string }

    # ===== 费用变更 =====
...
    ProjectFileListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ProjectFile' }
    ProjectFileResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/ProjectFile' }
    FileMetadataResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/FileMetadata' }
    ThesisVersionResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/ThesisVersion' }
    ThesisVersionListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ThesisVersion' }
...
  responses:
...
    NotFoundError:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
paths:
...
  /api/projects/{id}/files:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      operationId: projectsListFiles
      summary: 列出项目附件（sample_doc / source_code）
      tags: [files]
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [sample_doc, source_code]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectFileListResponse' }
    post:
      operationId: projectsAttachFile
      summary: 把已上传的 file 挂到项目下指定 category（用户上传源码 / 参考样稿后调用）
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileId, category]
              properties:
                fileId:
                  type: integer
                  format: int64
                category:
                  type: string
                  enum: [sample_doc, source_code]
                remark:
                  type: string
                  description: "可选备注：源码版本号 / 上传说明 / 样稿描述等（migration 0013 project_files.remark）"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectFileResponse' }
...
  /api/projects/{id}/thesis-versions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      operationId: projectsListThesisVersions
      tags: [thesis]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThesisVersionListResponse' }
    post:
      operationId: projectsCreateThesisVersion
      summary: 上传论文新版（永不覆盖，version_no 自动递增）
      tags: [thesis]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ThesisVersionCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThesisVersionResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ===== Files (multipart upload，文件本身用 multipart，元信息走 envelope) =====
  /api/files:
    post:
      operationId: filesUpload
      summary: 上传文件，返回文件元数据
      tags: [files]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: 文件已上传
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileMetadataResponse' }
        '413':
          description: 文件过大
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '415':
          description: MIME 不允许
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /api/files/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      operationId: filesDownload
      summary: 下载文件（流式）
      tags: [files]
      responses:
        '200':
          description: 文件内容
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404': { $ref: '#/components/responses/NotFoundError' }

  # ===== Notifications =====
...